backend/requirements.txt
fastapi
uvicorn
xgboost
pandas
scikit-learn
numpy
pillow
requests
python-multipart
torch
transformers
